pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "dexscreener-lead-bot"
version = "1.0.0"
description = "Telegram bot that monitors Dexscreener for new token pairs and extracts founder contact intelligence"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-telegram-bot>=21.0",
    "telethon>=1.36.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "tenacity>=8.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
